+++ src/components/shared/DonationBox.js
@@ -1,5 +1,5 @@
 import { colors, typography } from '../../themes/common-variables'
-import { screen } from '../../themes/screen'
+import mq from '../../utils/media-query'
 import DonationLink from '@twreporter/react-components/lib/donation-link-with-utm'
 import React, { PureComponent } from 'react'
 import styled from 'styled-components'
@@ -15,7 +15,7 @@ const Container = styled.div`
   min-height: 284px;
   background: ${colors.white};
   border-left: solid 1px ${colors.secondaryColor};
-  ${screen.mobile`
+  ${mq.mobileOnly`
     margin: 40px auto;
     width: 320px;
     min-height: 335px;
@@ -31,7 +31,7 @@ const Title = styled.p`
   font-weight: ${typography.font.weight.bold};
   color: ${colors.black};
   margin-bottom: 15px;
-  ${screen.mobile`
+  ${mq.mobileOnly`
     margin-bottom: 18px;
   `}
 `
@@ -46,7 +46,7 @@ const Donate = styled.div`
   width: 100%;
   height: 55px;
   margin-top: 50px;
-  ${screen.mobile`
+  ${mq.mobileOnly`
     margin-top: 40px;
   `}
   a{