+++ src/constants/release-branch.js
@@ -1,7 +1,0 @@
-export default {
-  master: 'master',
-  test: 'test',
-  staging: 'staging',
-  preview: 'preview',
-  release: 'release'
-}